requirements.txt
bentoml
torchvision
joblib
tqdm
pip-chill
wincertstore
torchsummary
opencv-python
from_root
gdown
pydrive
-e .
